backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
reportlab>=4.2.0
groq>=0.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
email-validator>=2.1.1
